+++ Views/GeomembraneForm.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Data.Entity;
 using Mine.DataModel;
+using MetroFramework;
 
 namespace Mine.Views
 {
@@ -60,23 +61,41 @@ namespace Mine.Views
 
         private void BtnAdd_Click(object sender, EventArgs e)
         {
-            GeoMembrane gm = new GeoMembrane()
+            try
             {
-                ProjectDevId=pdId,
-                SeamingDate=Convert.ToDateTime(DateDev.Text),
-                SeamNo=TxtSeamNo.Text,
-                SeamTime=TxtSeamTime.Text,
-                WedgeTemp=Convert.ToDouble(TxtTemp.Text),
-                WedgeSpeed=Convert.ToDouble(TxtSpeed.Text),
-                SeamLength=Convert.ToDouble(TxtSeamLength.Text),
-                CarryOver=Convert.ToDouble(TxtCarryOver.Text),
-                Destructive=TxtDestructive.Text,
-                RemarksDestructive=TxtRemarks.Text
-            };
+                GeoMembrane gm = new GeoMembrane()
+                {
+                    ProjectDevId = pdId,
+                    SeamingDate = Convert.ToDateTime(DateDev.Text),
+                    SeamNo = TxtSeamNo.Text,
+                    SeamTime = TxtSeamTime.Text,
+                    WedgeTemp = Convert.ToDouble(TxtTemp.Text),
+                    WedgeSpeed = Convert.ToDouble(TxtSpeed.Text),
+                    SeamLength = Convert.ToDouble(TxtSeamLength.Text),
+                    CarryOver = Convert.ToDouble(TxtCarryOver.Text),
+                    Destructive = TxtDestructive.Text,
+                    RemarksDestructive = TxtRemarks.Text
+                };
 
-            db.GeoMembranes.Add(gm);
-            db.SaveChanges();
-            GeomembraneForm_Load(sender, e);
+                db.GeoMembranes.Add(gm);
+                db.SaveChanges();
+                
+                MetroMessageBox.Show(this, "Saved successfully", "Info", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                GeomembraneForm_Load(sender, e);
+                TxtSeamNo.Text = "";
+                TxtSeamTime.Text = "";
+                TxtTemp.Text = "";
+                TxtSpeed.Text = "";
+                TxtSeamLength.Text = "";
+                TxtCarryOver.Text = "";
+                TxtDestructive.Text = "";
+                TxtRemarks.Text = "";
+            }
+            catch (Exception ex)
+            {
+                MetroMessageBox.Show(this, "Error"+ex.ToString(), "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+           
         }
 
         private void BtnEdit_Click(object sender, EventArgs e)
@@ -97,29 +116,56 @@ namespace Mine.Views
 
         private void BtnDelete_Click(object sender, EventArgs e)
         {
-            int pdId = Convert.ToInt32(DGVGeo.CurrentRow.Cells[0].Value);
-            var geo=db.GeoMembranes.Find(pdId);
-            db.GeoMembranes.Remove(geo);
-            db.SaveChanges();
-            GeomembraneForm_Load(sender,e);
-
+            try
+            {
+                int pdId = Convert.ToInt32(DGVGeo.CurrentRow.Cells[0].Value);
+                var geo = db.GeoMembranes.Find(pdId);
+                if (MetroFramework.MetroMessageBox.Show(this, "Are you sure want to delete this record?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+                {
+                    db.GeoMembranes.Remove(geo);
+                    db.SaveChanges();
+                    GeomembraneForm_Load(sender, e);
+                }
+               
+            }
+            catch (Exception ex)
+            {
+                MetroMessageBox.Show(this, "Error in your record " + ex.ToString(), "Info", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
 
 
         }
 
         private void BtnSave_Click(object sender, EventArgs e)
         {
-            var id=Convert.ToInt32(TxtGeoId.Text);
-            var geo = db.GeoMembranes.Find(id);
-            geo.Destructive = TxtDestructive.Text;
-            geo.RemarksDestructive = TxtRemarks.Text;
-            geo.SeamLength = Convert.ToDouble(TxtSeamLength.Text);
-            geo.SeamNo = TxtSeamNo.Text;
-            geo.SeamTime = TxtSeamTime.Text;
-            geo.WedgeSpeed = Convert.ToDouble(TxtSpeed.Text);
-            geo.WedgeTemp = Convert.ToDouble(TxtTemp.Text);
-            db.SaveChanges();
-            GeomembraneForm_Load(sender, e) ;
+            try
+            {
+                var id = Convert.ToInt32(TxtGeoId.Text);
+                var geo = db.GeoMembranes.Find(id);
+                geo.Destructive = TxtDestructive.Text;
+                geo.RemarksDestructive = TxtRemarks.Text;
+                geo.SeamLength = Convert.ToDouble(TxtSeamLength.Text);
+                geo.SeamNo = TxtSeamNo.Text;
+                geo.SeamTime = TxtSeamTime.Text;
+                geo.WedgeSpeed = Convert.ToDouble(TxtSpeed.Text);
+                geo.WedgeTemp = Convert.ToDouble(TxtTemp.Text);
+                db.SaveChanges();
+                GeomembraneForm_Load(sender, e);
+                MetroMessageBox.Show(this, "Updated successfully", "Info", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                TxtDestructive.Text=string.Empty;
+                TxtRemarks.Text=string.Empty;
+                TxtSeamLength.Text=string.Empty;
+                TxtSeamNo.Text=string.Empty;
+                TxtSeamTime.Text=string.Empty;
+                TxtSpeed.Text=string.Empty;
+                TxtTemp.Text=string.Empty;
+            }
+            catch (Exception ex)
+            {
+
+                MetroMessageBox.Show(this, "Error in your record " + ex.ToString(), "Info", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            
         }
 
         private void BtnPdf_Click(object sender, EventArgs e)
